Checksum, Downloads, and Unzips Quicklet Manifests
This project includes a feature-complete Quicklet client manifest
downloader. It will download an ini file from the download server,
verify the directory contains files with the same checksum, and
should files differ, initiate a manifest download (and checksum
the zip file and unzip).  This also includes a progress bar for
downloads (I thought unzipping/checksumming are fast enough to not
require a progress dialog so those can happen in the background).

Sorry for the long wait--it's finally here. Thanks @ophiesay
Fixes #3. Updates for #4.